Moved functions from main.py to main_parsers.py, Fixed import error
